# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This api is for testing the swagger ui",
    "version": "1.0.7",
    "title": "Express API",
    "contact": {
      "email": "test@gmail.com"
    }
  },
  "host": "0.0.0.0:5000",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/users/me/avatar": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Upload avatar",
        "description": "This can only be done by logged in user.",
        "operationId": "uploadAvatar",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "consumes": [
          "image/jpeg",
          "image/png",
          "image/webp"
        ],
        "produces": [
          "application/json",
          "application/xml"
        ],
        "responses": {
          "201": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/PrivateUserResponse"
            }
          },
          "400": {
            "description": "Avtar must be a single file"
          },
          "413": {
            "description": "Avatar has to be less than 5 mb"
          },
          "415": {
            "description": "Avatar has invalid file extension"
          }
        }
      },
      "delete": {
        "tags": [
          "user"
        ],
        "summary": "Delete avatar",
        "description": "This can only be done by logged in user.",
        "operationId": "deleteAvatar",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "successful operation"
          },
          "401": {
            "description": "Invalid token/No access token provided"
          }
        }
      }
    },
...
    "/auth/verify-email": {
      "put": {
        "tags": [
          "auth"
        ],
        "summary": "Verify email",
        "description": "This can only be done by registered user.",
        "operationId": "verifyEmail",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "successful operation"
          },
          "401": {
            "description": "Invalid token/No action token provided"
          }
        }
      }
    }
  },
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization"
    }
  },
  "definitions": {
...
    "PrivateUserResponse": {
      "type": "object",
      "required": [
        "_id",
        "name",
        "email",
        "role",
        "isDeleted",
        "isVerified"
      ],
      "properties": {
        "_id": {
          "type": "string",
          "format": "uuid",
          "description": "The unique identifier for a user"
        },
        "name": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "age": {
          "type": "number"
        },
        "phone": {
          "type": "string"
        },
        "avatar": {
          "type": "string"
        },
        "role": {
          "type": "string",
          "enum": [
            "admin",
            "user"
          ]
        },
        "isDeleted": {
          "type": "boolean"
        },
        "isVerified": {
          "type": "boolean"
        }
      }
    },
...
  }
}